Tools for examining modules on known repositories

When exploring existing modules, these tools help identify and navigate modules and their interdependencies.

listModules(
  keywords,
  accounts,
  omit = c("fireSense_dataPrepFitRas"),
  purge = FALSE,
  verbose = getOption("Require.verbose", 1L)
)

moduleDependencies(
  modules,
  modulePath = getOption("reproducible.modulePath", ".")
)

moduleDependenciesToGraph(md)

PlotModuleGraph(graph)

Arguments

keywords

A vector of character strings that will be used as keywords for identify modules

accounts

A vector of character strings identifying GitHub accounts e.g., PredictiveEcology to search.

omit

A vector of character strings of repositories to ignore.

purge

There is some internal caching that occurs. Setting this to TRUE will remove any cached data that is part of the requested accounts and keywords.

verbose

Numeric or logical indicating how verbose should the function be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE, then minimal outputs; if 1 or TRUE, more outputs; 2 even more. NOTE: in Require function, when verbose >= 2, the return object will have an attribute: attr(.., "Require") which has lots of information about the processes of the installs.

modules

A named list of character strings of full module names, as returned by listModules.

modulePath

A character string indicating the path where the modules are located.

md

A data.table with columns from and to, showing relationships of objects in modules. Likely from moduleDependencies.

graph

An igraph object to plot. Likely returned by moduleDependenciesToGraph.

Value

listModules returns a named list of all SpaDES modules in the given repositories with the accounts and keywords provided.

See also

Examples

listModules(accounts = "PredictiveEcology", "none")
#> searching keyword: none in PredictiveEcology
#> $PredictiveEcology
#> character(0)
#>