Make DESCRIPTION file(s) from SpaDES module metadata

Make DESCRIPTION file(s) from SpaDES module metadata

makeDESCRIPTIONproject(
  modules,
  modulePath,
  projectPath = ".",
  singleDESCRIPTION = TRUE,
  package = "Project",
  title = "Project",
  description = "Project",
  version = "1.0.0",
  authors = Sys.info()["user"],
  write = TRUE,
  verbose = getOption("Require.verbose")
)

makeDESCRIPTION(
  modules,
  modulePath,
  projectPath = ".",
  singleDESCRIPTION = FALSE,
  package,
  title,
  date,
  description,
  version,
  authors,
  write = TRUE,
  verbose,
  metadataList,
  ...
)

Arguments

modules

A character vector of module names

modulePath

Character. The path with modules, usually modulePath() or paths$modulePath

projectPath

Character. Only used if singleDESCRIPTION = TRUE

singleDESCRIPTION

Logical. If TRUE, there be only one DESCRIPTION file written for all modules, i.e., all reqdPkgs will be trimmed for redundancies and put into the single project-level DESCRIPTION file.

package

The name inserted into the "Package" entry in DESCRIPTION

title

The string inserted into the "Title" entry in DESCRIPTION

description

The string inserted into the "Description" entry in DESCRIPTION

version

The string inserted into the "Version" entry in DESCRIPTION

authors

The string inserted into the "Authors" entry in DESCRIPTION

write

Logical. If TRUE, then it will write the DESCRIPTION file either in the modulePath (if singleDESCRIPTION = FALSE) or projectPath (if singleDESCRIPTION = TRUE)

verbose

Numeric or logical indicating how verbose should the function be. If -1 or -2, then as little verbosity as possible. If 0 or FALSE, then minimal outputs; if 1 or TRUE, more outputs; 2 even more. NOTE: in Require function, when verbose >= 2, also returns details as if returnDetails = TRUE (for backwards compatibility).

date

Date to enter into DESCRIPTION file. Defaults to Sys.Date()

metadataList

The parsed source code from a module. Must include defineModule metadata.

...

Currently not used.